+++ taotao-cloud-microservice/taotao-cloud-distribution/taotao-cloud-distribution-biz/src/main/java/com/taotao/cloud/distribution/biz/service/DistributionGoodsService.java
@@ -78,6 +78,6 @@ public interface DistributionGoodsService extends IService<DistributionGoods> {
      * @param storeId 店铺id
      * @return
      */
-    DistributionGoods checked(String skuId, Double commission, String storeId);
+    DistributionGoods checked(String skuId, BigDecimal commission, String storeId);
 
 }
