+++ taotao-cloud-microservice/taotao-cloud-order/taotao-cloud-order-api/src/main/java/com/taotao/cloud/order/api/vo/cart/CartVO.java
@@ -60,7 +60,7 @@ public class CartVO extends CartBase implements Serializable {
 	private Integer giftPoint;
 
 	@Schema(description = "重量")
-	private Double weight;
+	private BigDecimal weight;
 
 	@Schema(description = "购物车商品数量")
 	private Integer goodsNum;
